pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-bot"
version = "0.1.0"
description = "BTC funding carry bot using Binance spot and USDⓈ-M via ccxt"
authors = [{name = "Codex", email = "codex@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ccxt>=4.0.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.3.0",
    "black>=23.12.0",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E","F","I","B","UP","N"
]
ignore = [
    "E501"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-s"
pythonpath = ["src"]
